+++ R/learner_RWeka_regr_M5Rules.R
@@ -37,8 +37,6 @@ LearnerRegrM5Rules = R6Class("LearnerRegrM5Rules",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
 
-      warning("Deprecated, please use `lrn(regr.rweka, model = 'M5Rules')` in the future.")
-
       ps = ps(
           subset = p_uty(tags = c("train", "pars")),
           na.action = p_uty(tags = c("train", "pars")),