+++ R/zzz.R
@@ -1,3 +1,4 @@
+#' @import checkmate
 #' @import paradox
 #' @import mlr3misc
 #' @importFrom R6 R6Class